refactor: put files in the right folders
