--- frontend/src/components/organisms/index.ts
@@ -1,13 +1,0 @@
-import Aerodynamics from "./Aerodynamics";
-import Performance from "./Performance";
-import Settings from "./Settings";
-import Turn from "./Turn";
-import Weight from "./Weight";
-
-export {
-    Aerodynamics,
-    Performance,
-    Settings,
-    Turn,
-    Weight
-}
