+++ frontend/src/features/aerodynamics/Aerodynamics.tsx
