+++ frontend/src/features/weight_distribution/Weight.tsx
